feat: move hook to systems core
