+++ packages/app/src/App.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Providers } from '~/systems/Core';
 
 import { IS_DEVELOPMENT, IS_TEST } from './config';
-import { useRecoverWelcomeFromError } from './hooks';
 import { getRoutes } from './routes';
+import { useRecoverWelcomeFromError } from './systems/Core/hooks/useRecoverWelcomeFromError';
 
 const ThrowError = React.lazy(
   () => import('./systems/Error/components/ThrowError')
